changed prefab structure for voice chat and connected lip sync to it
