--- Assets/MirageXR/Networking/Scripts/LipSyncInitializationBridge.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 26ced2f01adc8cf4babc9730b9ae2d43
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
